move estimators to transformations.compose
