+++ sktime/transformations/series/compose.py
@@ -8,11 +8,11 @@ __all__ = ["OptionalPassthrough", "ColumnwiseTransformer", "YtoX"]
 
 from warnings import warn
 
-import pandas as pd
-from sklearn.utils.metaestimators import if_delegate_has_method
-
-from sktime.transformations.base import BaseTransformer
-from sktime.utils.validation.series import check_series
+from sktime.transformations.compose import (
+    ColumnwiseTransformer,
+    OptionalPassthrough,
+    YtoX,
+)
 
 warn(
     "transformations.series.compose is deprecated and will be removed in"
@@ -20,497 +20,3 @@ warn(
     "Please change imports to transformations.compose to avoid breakage.",
     DeprecationWarning,
 )
-
-
-# mtypes for Series, Panel, Hierarchical,
-# with exception of some ambiguous and discouraged mtypes
-CORE_MTYPES = [
-    "pd.DataFrame",
-    "np.ndarray",
-    "pd.Series",
-    "pd-multiindex",
-    "df-list",
-    "nested_univ",
-    "numpy3D",
-    "pd_multiindex_hier",
-]
-
-
-class OptionalPassthrough(BaseTransformer):
-    """Wrap an existing transformer to tune whether to include it in a pipeline.
-
-    Allows tuning the implicit hyperparameter whether or not to use a
-    particular transformer inside a pipeline (e.g. TranformedTargetForecaster)
-    or not. This is achieved by the hyperparameter `passthrough`
-    which can be added to a tuning grid then (see example).
-
-    Parameters
-    ----------
-    transformer : Estimator
-        scikit-learn-like or sktime-like transformer to fit and apply to series.
-    passthrough : bool, default=False
-       Whether to apply the given transformer or to just
-        passthrough the data (identity transformation). If, True the transformer
-        is not applied and the OptionalPassthrough uses the identity
-        transformation.
-
-    Examples
-    --------
-    >>> from sktime.datasets import load_airline
-    >>> from sktime.forecasting.naive import NaiveForecaster
-    >>> from sktime.transformations.series.compose import OptionalPassthrough
-    >>> from sktime.transformations.series.detrend import Deseasonalizer
-    >>> from sktime.transformations.series.adapt import TabularToSeriesAdaptor
-    >>> from sktime.forecasting.compose import TransformedTargetForecaster
-    >>> from sktime.forecasting.model_selection import (
-    ...     ForecastingGridSearchCV,
-    ...     SlidingWindowSplitter)
-    >>> from sklearn.preprocessing import StandardScaler
-    >>> # create pipeline
-    >>> pipe = TransformedTargetForecaster(steps=[
-    ...     ("deseasonalizer", OptionalPassthrough(Deseasonalizer())),
-    ...     ("scaler", OptionalPassthrough(TabularToSeriesAdaptor(StandardScaler()))),
-    ...     ("forecaster", NaiveForecaster())])
-    >>> # putting it all together in a grid search
-    >>> cv = SlidingWindowSplitter(
-    ...     initial_window=60,
-    ...     window_length=24,
-    ...     start_with_window=True,
-    ...     step_length=48)
-    >>> param_grid = {
-    ...     "deseasonalizer__passthrough" : [True, False],
-    ...     "scaler__transformer__transformer__with_mean": [True, False],
-    ...     "scaler__passthrough" : [True, False],
-    ...     "forecaster__strategy": ["drift", "mean", "last"]}
-    >>> gscv = ForecastingGridSearchCV(
-    ...     forecaster=pipe,
-    ...     param_grid=param_grid,
-    ...     cv=cv,
-    ...     n_jobs=-1)
-    >>> gscv_fitted = gscv.fit(load_airline())
-    """
-
-    _tags = {
-        "scitype:transform-input": "Series",
-        # what is the scitype of X: Series, or Panel
-        "scitype:transform-output": "Series",
-        # what scitype is returned: Primitives, Series, Panel
-        "scitype:instancewise": True,  # is this an instance-wise transform?
-        "X_inner_mtype": CORE_MTYPES,
-        # which mtypes do _fit/_predict support for X?
-        "y_inner_mtype": "None",  # which mtypes do _fit/_predict support for y?
-        "univariate-only": False,
-        "fit_is_empty": False,
-        "capability:inverse_transform": True,
-    }
-
-    def __init__(self, transformer, passthrough=False):
-        self.transformer = transformer
-        self.transformer_ = None
-        self.passthrough = passthrough
-        self._is_fitted = False
-        super(OptionalPassthrough, self).__init__()
-
-        # should be all tags, but not fit_is_empty
-        #   (_fit should not be skipped)
-        tags_to_clone = [
-            "scitype:transform-input",
-            "scitype:transform-output",
-            "scitype:instancewise",
-            "X_inner_mtype",
-            "y_inner_mtype",
-            "capability:inverse_transform",
-            "handles-missing-data",
-            "X-y-must-have-same-index",
-            "transform-returns-same-time-index",
-            "skip-inverse-transform",
-        ]
-        self.clone_tags(transformer, tag_names=tags_to_clone)
-
-    def _fit(self, X, y=None):
-        """Fit transformer to X and y.
-
-        private _fit containing the core logic, called from fit
-
-        Parameters
-        ----------
-        X : Series or Panel of mtype X_inner_mtype
-            if X_inner_mtype is list, _fit must support all types in it
-            Data to fit transform to
-        y : Series or Panel of mtype y_inner_mtype, default=None
-            Additional data, e.g., labels for tarnsformation
-
-        Returns
-        -------
-        self: a fitted instance of the estimator
-        """
-        if not self.passthrough:
-            self.transformer_ = self.transformer.clone()
-            self.transformer_._fit(X, y)
-        return self
-
-    def _transform(self, X, y=None):
-        """Transform X and return a transformed version.
-
-        private _transform containing the core logic, called from transform
-
-        Parameters
-        ----------
-        X : Series or Panel of mtype X_inner_mtype
-            if X_inner_mtype is list, _transform must support all types in it
-            Data to be transformed
-        y : Series or Panel of mtype y_inner_mtype, default=None
-            Additional data, e.g., labels for transformation
-
-        Returns
-        -------
-        transformed version of X
-        """
-        if not self.passthrough:
-            X = self.transformer_._transform(X, y)
-        return X
-
-    def _inverse_transform(self, X, y=None):
-        """Inverse transform, inverse operation to transform.
-
-        core logic
-
-        Parameters
-        ----------
-        X : Series or Panel of mtype X_inner_mtype
-            if X_inner_mtype is list, _inverse_transform must support all types in it
-            Data to be inverse transformed
-        y : Series or Panel of mtype y_inner_mtype, optional (default=None)
-            Additional data, e.g., labels for transformation
-
-        Returns
-        -------
-        inverse transformed version of X
-        """
-        if not self.passthrough:
-            X = self.transformer_._inverse_transform(X, y)
-        return X
-
-    @classmethod
-    def get_test_params(cls, parameter_set="default"):
-        """Return testing parameter settings for the estimator.
-
-        Parameters
-        ----------
-        parameter_set : str, default="default"
-            Name of the set of test parameters to return, for use in tests. If no
-            special parameters are defined for a value, will return `"default"` set.
-
-
-        Returns
-        -------
-        params : dict or list of dict, default = {}
-            Parameters to create testing instances of the class
-            Each dict are parameters to construct an "interesting" test instance, i.e.,
-            `MyClass(**params)` or `MyClass(**params[i])` creates a valid test instance.
-            `create_test_instance` uses the first (or only) dictionary in `params`
-        """
-        from sktime.transformations.series.boxcox import BoxCoxTransformer
-
-        return {"transformer": BoxCoxTransformer(), "passthrough": False}
-
-
-class ColumnwiseTransformer(BaseTransformer):
-    """Apply a transformer columnwise to multivariate series.
-
-    Overview: input multivariate time series and the transformer passed
-    in `transformer` parameter is applied to specified `columns`, each
-    column is handled as a univariate series. The resulting transformed
-    data has the same shape as input data.
-
-    Parameters
-    ----------
-    transformer : Estimator
-        scikit-learn-like or sktime-like transformer to fit and apply to series.
-    columns : list of str or None
-            Names of columns that are supposed to be transformed.
-            If None, all columns are transformed.
-
-    Attributes
-    ----------
-    transformers_ : dict of {str : transformer}
-        Maps columns to transformers.
-    columns_ : list of str
-        Names of columns that are supposed to be transformed.
-
-    See Also
-    --------
-    OptionalPassthrough
-
-    Examples
-    --------
-    >>> from sktime.datasets import load_longley
-    >>> from sktime.transformations.series.detrend import Detrender
-    >>> from sktime.transformations.series.compose import ColumnwiseTransformer
-    >>> _, X = load_longley()
-    >>> transformer = ColumnwiseTransformer(Detrender())
-    >>> Xt = transformer.fit_transform(X)
-    """
-
-    _tags = {
-        "scitype:transform-input": "Series",
-        # what is the scitype of X: Series, or Panel
-        "scitype:transform-output": "Series",
-        # what scitype is returned: Primitives, Series, Panel
-        "scitype:instancewise": True,  # is this an instance-wise transform?
-        "X_inner_mtype": "pd.DataFrame",
-        # which mtypes do _fit/_predict support for X?
-        "y_inner_mtype": "None",  # which mtypes do _fit/_predict support for y?
-        "univariate-only": False,
-        "fit_is_empty": False,
-    }
-
-    def __init__(self, transformer, columns=None):
-        self.transformer = transformer
-        self.columns = columns
-        super(ColumnwiseTransformer, self).__init__()
-
-        tags_to_clone = [
-            "y_inner_mtype",
-            "capability:inverse_transform",
-            "handles-missing-data",
-            "X-y-must-have-same-index",
-            "transform-returns-same-time-index",
-            "skip-inverse-transform",
-        ]
-        self.clone_tags(transformer, tag_names=tags_to_clone)
-
-    def _fit(self, X, y=None):
-        """Fit transformer to X and y.
-
-        private _fit containing the core logic, called from fit
-
-        Parameters
-        ----------
-        X : pd.DataFrame
-            Data to fit transform to
-        y : Series or Panel, default=None
-            Additional data, e.g., labels for transformation
-
-        Returns
-        -------
-        self: a fitted instance of the estimator
-        """
-        # check that columns are None or list of strings
-        if self.columns is not None:
-            if not isinstance(self.columns, list) and all(
-                isinstance(s, str) for s in self.columns
-            ):
-                raise ValueError("Columns need to be a list of strings or None.")
-
-        # set self.columns_ to columns that are going to be transformed
-        # (all if self.columns is None)
-        self.columns_ = self.columns
-        if self.columns_ is None:
-            self.columns_ = X.columns
-
-        # make sure z contains all columns that the user wants to transform
-        _check_columns(X, selected_columns=self.columns_)
-
-        # fit by iterating over columns
-        self.transformers_ = {}
-        for colname in self.columns_:
-            transformer = self.transformer.clone()
-            self.transformers_[colname] = transformer
-            self.transformers_[colname].fit(X[colname], y)
-        return self
-
-    def _transform(self, X, y=None):
-        """Transform X and return a transformed version.
-
-        private _transform containing the core logic, called from transform
-
-        Returns a transformed version of X by iterating over specified
-        columns and applying the wrapped transformer to them.
-
-        Parameters
-        ----------
-        X : pd.DataFrame
-            Data to be transformed
-        y : Series or Panel, default=None
-            Additional data, e.g., labels for transformation
-
-        Returns
-        -------
-        Xt : pd.DataFrame
-            transformed version of X
-        """
-        # make copy of z
-        X = X.copy()
-
-        # make sure z contains all columns that the user wants to transform
-        _check_columns(X, selected_columns=self.columns_)
-        for colname in self.columns_:
-            X[colname] = self.transformers_[colname].transform(X[colname], y)
-        return X
-
-    def _inverse_transform(self, X, y=None):
-        """Logic used by `inverse_transform` to reverse transformation on `X`.
-
-        Returns an inverse-transformed version of X by iterating over specified
-        columns and applying the univariate series transformer to them.
-        Only works if `self.transformer` has an `inverse_transform` method.
-
-        Parameters
-        ----------
-        X : pd.DataFrame
-            Data to be inverse transformed
-        y : Series or Panel, default=None
-            Additional data, e.g., labels for transformation
-
-        Returns
-        -------
-        Xt : pd.DataFrame
-            inverse transformed version of X
-        """
-        # make copy of z
-        X = X.copy()
-
-        # make sure z contains all columns that the user wants to transform
-        _check_columns(X, selected_columns=self.columns_)
-
-        # iterate over columns that are supposed to be inverse_transformed
-        for colname in self.columns_:
-            X[colname] = self.transformers_[colname].inverse_transform(X[colname], y)
-
-        return X
-
-    @if_delegate_has_method(delegate="transformer")
-    def update(self, X, y=None, update_params=True):
-        """Update parameters.
-
-        Update the parameters of the estimator with new data
-        by iterating over specified columns.
-        Only works if `self.transformer` has an `update` method.
-
-        Parameters
-        ----------
-        X : pd.Series
-            New time series.
-        update_params : bool, optional, default=True
-
-        Returns
-        -------
-        self : an instance of self
-        """
-        z = check_series(X)
-
-        # make z a pd.DataFrame in univariate case
-        if isinstance(z, pd.Series):
-            z = z.to_frame()
-
-        # make sure z contains all columns that the user wants to transform
-        _check_columns(z, selected_columns=self.columns_)
-        for colname in self.columns_:
-            self.transformers_[colname].update(z[colname], X)
-        return self
-
-    @classmethod
-    def get_test_params(cls, parameter_set="default"):
-        """Return testing parameter settings for the estimator.
-
-        Parameters
-        ----------
-        parameter_set : str, default="default"
-            Name of the set of test parameters to return, for use in tests. If no
-            special parameters are defined for a value, will return `"default"` set.
-
-
-        Returns
-        -------
-        params : dict or list of dict, default = {}
-            Parameters to create testing instances of the class
-            Each dict are parameters to construct an "interesting" test instance, i.e.,
-            `MyClass(**params)` or `MyClass(**params[i])` creates a valid test instance.
-            `create_test_instance` uses the first (or only) dictionary in `params`
-        """
-        from sktime.transformations.series.detrend import Detrender
-
-        return {"transformer": Detrender()}
-
-
-def _check_columns(z, selected_columns):
-    # make sure z contains all columns that the user wants to transform
-    z_wanted_keys = set(selected_columns)
-    z_new_keys = set(z.columns)
-    difference = z_wanted_keys.difference(z_new_keys)
-    if len(difference) != 0:
-        raise ValueError("Missing columns" + str(difference) + "in Z.")
-
-
-def _check_is_pdseries(z):
-    # make z a pd.Dataframe in univariate case
-    is_series = False
-    if isinstance(z, pd.Series):
-        z = z.to_frame()
-        is_series = True
-    return z, is_series
-
-
-class YtoX(BaseTransformer):
-    """Create exogeneous features which are a copy of the endogenous data.
-
-    Replaces exogeneous features (`X`) by endogeneous data (`y`).
-
-    To *add* instead of *replace*, use `FeatureUnion`.
-
-    Parameters
-    ----------
-    no parameters
-    """
-
-    _tags = {
-        "transform-returns-same-time-index": True,
-        "skip-inverse-transform": False,
-        "univariate-only": False,
-        "X_inner_mtype": ["pd.DataFrame", "pd-multiindex", "pd_multiindex_hier"],
-        "y_inner_mtype": ["pd.DataFrame", "pd-multiindex", "pd_multiindex_hier"],
-        "scitype:y": "both",
-        "fit_is_empty": True,
-        "requires_y": True,
-    }
-
-    def __init__(self):
-        super(YtoX, self).__init__()
-
-    def _transform(self, X, y=None):
-        """Transform X and return a transformed version.
-
-        private _transform containing core logic, called from transform
-
-        Parameters
-        ----------
-        X : time series or panel in one of the pd.DataFrame formats
-            Data to be transformed
-        y : time series or panel in one of the pd.DataFrame formats
-            Additional data, e.g., labels for transformation
-
-        Returns
-        -------
-        y, as a transformed version of X
-        """
-        return y
-
-    def _inverse_transform(self, X, y=None):
-        """Inverse transform, inverse operation to transform.
-
-        Drops featurized column that was added in transform().
-
-        Parameters
-        ----------
-        X : Series or Panel of mtype X_inner_mtype
-            if X_inner_mtype is list, _inverse_transform must support all types in it
-            Data to be inverse transformed
-        y : Series or Panel of mtype y_inner_mtype, optional (default=None)
-            Additional data, e.g., labels for transformation
-
-        Returns
-        -------
-        inverse transformed version of X
-        """
-        return y
